Added translation for names and comments into 5 languages
