+++ portfolio/src/main/webapp/education.html
@@ -6,7 +6,7 @@
   <title>Education</title>
   <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@500&display=swap" rel="stylesheet">
   <link rel="stylesheet" href="style.css">
-  <script src="script.js"></script>
+  <script src="education.js"></script>
 </head>
 
 <body id="content">
