+++ portfolio/src/main/webapp/education.js
@@ -1,0 +1,27 @@
+// Copyright 2019 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Shows or hides the list of courses for a particular semester upon user click
+function toggleCourseList(listObject) {
+  const semester = listObject.id;
+  const semCourseList = "courselist" + semester;
+  const lst = document.getElementById(semCourseList);
+  if (listObject.className === "up") {
+    lst.className = "lstdown";
+    listObject.className = "down";
+  } else {
+    lst.className = "lstup";
+    listObject.className = "up";
+  }
+}
