+++ portfolio/src/main/webapp/projects.html
@@ -6,7 +6,6 @@
   <title>Projects</title>
   <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@500&display=swap" rel="stylesheet">
   <link rel="stylesheet" href="style.css">
-  <script src="script.js"></script>
 </head>
 
 <body id="content">
